Fix autotools detection of -Wno-* compiler flags
From gcc version 4.4, the manual states: "When an unrecognized warning option is requested
(e.g., -Wunknown-warning), GCC emits a diagnostic stating that the option is not recognized.
However, if the -Wno- form is used, the behavior is slightly different: no diagnostic is
produced for -Wno-unknown-warning unless other diagnostics are being produced."

This means that nlbuild-autotools won't detect invalid -Wno-* warning flags. As a
consequence, gcc is passed some invalid -Wno-* options.

This commit adds code to configure.ac that passes the positive version of each warning
flag (for example -Wno-gnu-zero-variadic-macro-arguments is checked as
-Wgnu-zero-variadic-macro-arguments), so the warning flag doesn't get added to
CFLAGS/CXXFLAGS).