[network-data] support registration on child devices
A device's request to register network data information is triggered
by a call to NetworkData::Local::SendServerDataNotification().
However, the call was only made by devices operating in the router
or leader roles.

This commit moves the call so that network data registration is
triggered in the child role, in addition to the router and leader
roles.