[build] Support building platform code without examples
- [build] Added --with-platform configure option to enable building the platform libraries
without building the example applications or any supporting libraries used solely by the
examples.

- [nrf528xx] Modified the nrf528xx-specific makefiles such that Nordic SDK code located in
third_party/NordicSemiconductor is only built/referenced when building the example
applications.  When building just the platform libraries, the caller of configure is
expected to provide the correct arguments to use an external copy of the Nordic nRF5 SDK.
An exception to this is the 802.15.4 radio driver, which is always built using the sources
in third_party/NordicSemiconductor/drivers/radio.

- [nrf528xx] Made various minor edits to #include statements in the nrf528xx platform code.
This was necessary because of the difference in directory structure between the code in
third_party/NordicSemiconductor and that in the formally published Nordic nRF5 SDK.