[posix] fix handling SPI interrupts
When SPI is already interrupted before select(), the wait timeout will
be set to 0, and select() will return 0 immediately if there are no new
events. As a result, cause an incorrect OT_ERROR_RESPONSE_TIMEOUT error.

This PR fixes this issue by check if data is ready before checking the
select() result.