[toolchain] Specify alignment of struct otExtAddress
The alignment of the address needs to be at a 64-bit (8-byte)
boundary, because the value is freely cast to uint64.
See for example the cast in the call to rfCoreFindExtSrcMatchIdx()
in otPlatRadioAddSrcMatchExtEntry() in examples/platforms/cc1352/radio.c.

Some macros are added to include/openthread/platform/toolchain.h to
make it possible to align structure members regardless of the compiler.

This helps fix compilation with arm-none-eabi-gcc 9.2.0.