Refactoring of POSIX netif code, mostly to accommodate BSDs
This code (likely a first attempt, as I'm sure there's some stuff in here that
folks will want to change) does the following:

- provides an implementation to manage networking notices from a PF_ROUTE socket,
	instead of (the Linux-specific) PF_NETLINK

- breaks out some of the platform-specific mechanisms for configuring a tunnel
	device (Linux, mac OS, and NetBSD are all different in some ways)

- provides handling of I/O to the tun drivers on BSDs (where there's a 4-byte
	header prepended to identify the IP version)

- adds some debugging info and more deliberate error handling (I suspect
	there are some more OpenThread-friendly ways to do some of the
	debugging, but I didn't try to figure them out)

I've used most of this code for several months on mac OS and NetBSD, and I
wanted to get a version of it shared for review.  I'd love to see it
merged in so that I don't have to maintain it completely by myself (and perhaps
others will want to use it on those platforms), but I suspect it'll need
some changes first.