[core/common] Add an option to move instance on heap
There is limit data segment on some FreeRTOS platform, because the
instance is very large, we may want to move the instance from static
data segment to dynamic allocated heap.

The patch introduced an new define option:

OPENTHREAD_CONFIG_INSTANCE_ON_HEAP

When it is defined, the instance will be allocated dynamically on heap.

Signed-off-by: Zang MingJie <mingjiez@google.com>