OPENTHREAD-434: Fix otPlatRadioGetRssi() for efr32
This fixes an issue with Channel Monitoring for the EFR32 platform (MG1, MG12, MG13), where in, turning the radio to a
new channel and immediately gathering background RSSI information was returning invalid results.

A further investigation with the RAIL team indicated that this issue was occurring because the usage of `RAIL_GetRssi()`
is incorrect. While the Rail_GetRssi API functions as documented, the problem is that the current implementation of
this API returns valid results only in certain specified radio states and does not fit the use case for the OpenThread
Channel Monitoring feature. (i.e computing RSSI value in the RxWarm radio state).

To fix the issue, I am switching the affected platforms to the efr32mg21 implementation for `otPlatRadioGetRssi()`,
which uses `RAIL_StartAverageRssi()` instead of `RAIL_GetRssi()`. `RAIL_StartAverageRssi()` uses hardware to trigger an
interrupt once the RSSI is valid and should provide valid results.

This approach was added by Marven when testing DMP support with OpenThread (https://github.com/openthread/openthread/pull/4321)